Make send-as-user more robust.
Add a List-Id field when sending as a user. Spam filters assume that something
is not a mailing list if it doesn't set the List-Id header, so we need to set
it when we pretend to be a mailing list.
Also, set the correct user name when sending as a user.